time: add an explicit type cast to tv_nsec
This line silently converts a double to a long. Evidently MSVC feels
something fishy about it.

> warning C4244: '=': conversion from 'double' to 'long', possible loss of data

I see this warnings more than 50 times during the compilation. Let's
just fix it.

+++ include/fluent-bit/flb_time.h
@@ -65,7 +65,7 @@ static inline void flb_time_copy(struct flb_time *dst, struct flb_time *src)
 static inline void flb_time_from_double(struct flb_time *dst, double d)
 {
     dst->tm.tv_sec = (int) d;
-    dst->tm.tv_nsec = (d - dst->tm.tv_sec) * 1000000000L;
+    dst->tm.tv_nsec = (long) ((d - dst->tm.tv_sec) * 1000000000L);
 }
 
 static inline int flb_time_equal(struct flb_time *t0, struct flb_time *t1) {
